Modularize execution modeling primitives
