+++ qmtl/sdk/execution_modeling/__init__.py
@@ -1,0 +1,40 @@
+"""Execution modeling primitives with reusable data models and helpers."""
+
+from .engine import ExecutionModel, create_market_data_from_ohlcv
+from .helpers import (
+    apply_latency,
+    base_slippage_from_bps,
+    calculate_commission,
+    calculate_market_impact,
+    calculate_slippage,
+    determine_fill_quantity,
+    spread_cost,
+)
+from .models import (
+    ExecutionFill,
+    MarketData,
+    OpenOrder,
+    OrderSide,
+    OrderStatus,
+    OrderType,
+    TimeInForce,
+)
+
+__all__ = [
+    "ExecutionModel",
+    "create_market_data_from_ohlcv",
+    "apply_latency",
+    "base_slippage_from_bps",
+    "calculate_commission",
+    "calculate_market_impact",
+    "calculate_slippage",
+    "determine_fill_quantity",
+    "spread_cost",
+    "ExecutionFill",
+    "MarketData",
+    "OpenOrder",
+    "OrderSide",
+    "OrderStatus",
+    "OrderType",
+    "TimeInForce",
+]
